[MISSING VTS MEASURE] data migration for '7225 9900 40'
https://www.trade-tariff.service.gov.uk/trade-tariff/commodities/7225990040#import

IN LIVE DB:
```7225990040: VTS validity_start_date: 2016-12-15 11:42:00 UTC - validity_end_date: 2016-12-15 15:04:00 UTC, geo area: ERGA OMNES, operation: update, operation_date: 2016-12-16
7225990040: VTS validity_start_date: 2016-12-15 15:04:00 UTC - validity_end_date: no date, geo area: ERGA OMNES, operation: update, operation_date: 2016-12-16
7225990040: VTS validity_start_date: 2017-01-01 00:00:00 UTC - validity_end_date: no date, geo area: ERGA OMNES, operation: update, operation_date: 2017-01-02
7225990040: VTS validity_start_date: 2017-04-01 00:00:00 UTC - validity_end_date: no date, geo area: ERGA OMNES, operation: update, operation_date: 2017-04-02
```

IN LOCAL DB:
```7225990040: VTS validity_start_date: 2016-12-15 11:42:00 UTC - validity_end_date: 2016-12-15 15:04:00 UTC, geo area: ERGA OMNES, operation: update, operation_date: 2016-12-16
7225990040: VTS validity_start_date: 2016-12-15 15:04:00 UTC - validity_end_date: no date, geo area: ERGA OMNES, operation: update, operation_date: 2016-12-16
7225990040: VTS validity_start_date: 2017-01-01 00:00:00 UTC - validity_end_date: no date, geo area: ERGA OMNES, operation: update, operation_date: 2017-01-02
7225990040: VTS validity_start_date: 2017-04-01 00:00:00 UTC - validity_end_date: no date, geo area: ERGA OMNES, operation: update, operation_date: 2017-04-02
```

IN CHIEF UPDATE FILES we have related actions:

create at *2016-12-16_KBT009(16351).txt*:
`103: "MFCM","15/12/2016:15:04:00","U","VT","S","B00",null,null,null,"15/12/2016:15:02:00","72259900 40","N","N","N"`
`104: "MFCM","15/12/2016:11:42:00","I","VT","S","B00",null,"15/12/2016:15:04:00",null,"15/12/2016:11:40:00","72259900 40","N","N","N"`

update at *2017-01-02_KBT009(17002).txt*:
`52237: "MFCM","01/01/2017:00:00:00","U","VT","S","B00",null,null,null,"21/12/2016:08:11:00","72259900 40","N","N","N"`

update at *2017-04-02_KBT009(17092).txt*:
`18128: "MFCM","01/04/2017:00:00:00","U","VT","S","B00",null,null,null,"28/03/2017:08:00:00","72259900 40","N","N","N"`

*Summary:* There are no any VTS measure in DB with validity_end_date in future for `7225990040` ,
that's why we no longer see it at https://www.trade-tariff.service.gov.uk/trade-tariff/commodities/7225990040#import
as validity_end_date expired at `28/03/2017:08:00:00`

[TRELLO STORY](https://trello.com/c/sGTRQBh4/457-tariff17-vat-is-missing-from-some-commodity-codes)
